Synchronise tide levels across all Shoals levels, so that high tide on one level implies high tide on all.

+++ crawl-ref/source/dgn-shoals.h
@@ -3,7 +3,7 @@
 
 void dgn_build_shoals_level(int level_number);
 void shoals_postprocess_level();
-void shoals_apply_tides(int turns_elapsed, bool force = false);
+void shoals_apply_tides(long turns_elapsed, bool force = false);
 void shoals_release_tide(monsters *caller);
 
 #ifdef WIZARD
